+++ alchemy-midlet/src/alchemy/fs/siemens/FS.java
@@ -126,7 +126,7 @@ public final class FS extends Filesystem implements Initable {
 		FileConnection fc = (FileConnection)Connector.open(root+file.path(), Connector.READ_WRITE);
 		try {
 			if (file.path().length() == 0) throw new SecurityException("Cannot delete root directory.");
-			fc.delete();
+			if (fc.exists()) fc.delete();
 		} finally {
 			fc.close();
 		}
